SIG dies immediately when if disconnects from dispatcher
This is a intermediate measure. Eventually, we want to have
proper reconnect functionality but until then it's better
to kill the process instead of keeping it alive and malfunctioning.
